CTP-4512 Revert e20650f (#92)
Previously we checked if block_portico_enrolments::get_modocc_mappings()
was present before calling this.  This was intended to avoid an error in
dev environments where Portico mappings were not present.  However,
since block_portico_enrolments must be installed to access the site
report, a dev environment without this block is of limited use.